+++ stream-chat-android-ui-common/src/main/kotlin/com/getstream/sdk/chat/viewmodel/messages/MessageListViewModel.kt
@@ -31,6 +31,8 @@ import io.getstream.chat.android.livedata.ChatDomain
 import io.getstream.chat.android.livedata.controller.ChannelController
 import io.getstream.chat.android.offline.experimental.channel.state.MessagesState
 import io.getstream.chat.android.offline.experimental.extensions.asReferenced
+import io.getstream.chat.android.offline.extensions.downloadAttachment
+import io.getstream.chat.android.offline.extensions.setMessageForReply
 import kotlinx.coroutines.flow.map
 import kotlin.properties.Delegates
 import io.getstream.chat.android.livedata.utils.Event as EventWrapper
@@ -351,14 +353,14 @@ public class MessageListViewModel @JvmOverloads constructor(
                 )
             }
             is Event.ReplyMessage -> {
-                domain.setMessageForReply(event.cid, event.repliedMessage).enqueue(
+                client.setMessageForReply(event.cid, event.repliedMessage).enqueue(
                     onError = { chatError ->
                         logger.logE("Could not reply message: ${chatError.message}. Cause: ${chatError.cause?.message}")
                     }
                 )
             }
             is Event.DownloadAttachment -> {
-                domain.downloadAttachment(event.attachment).enqueue(
+                client.downloadAttachment(event.attachment).enqueue(
                     onError = { chatError ->
                         logger.logE("Attachment download error: ${chatError.message}. Cause: ${chatError.cause?.message}")
                     }